:art: Migrating user related methods into a User class
