+++ README.md
@@ -21,27 +21,28 @@ try {
     $georide = new Client();
 
     // Create a new session ...
-    $georide->login('test@example.com', 'passw0rd');
+    $georide->user->login('test@example.com', 'passw0rd');
     $authToken = $georide->getAuthToken();
 
     // or resume an existing session
     $georide->setAuthToken($authToken);
 
     /**
-     * And then, you can call the API freely, 
+     * And then, you can call the API freely,
      * ex: get the trips from the first found tracker
      */
     $trackers = $georide->request('GET', '/user/trackers');
-    if (sizeof($trackers->trackers) === 0) {
+    if (sizeof($trackers) === 0) {
         throw new Exception("No tracker configured.");
     }
 
-    $tracker = $trackers->trackers[0];
+    $tracker = $trackers[0];
     $trips = $georide->request('GET', "/tracker/$tracker->trackerId/trips", [
         'from' => '2019-02-23T00:00:00Z',
         'to' => '2019-02-24T00:00:00Z'
     ]);
 
+    var_dump($trips);
 } catch (ApiException $e) {
     echo $e->getMessage();
 } catch (Exception $e) {
